use gate in post controller for checking access control

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use App\Http\Requests\PostRequest;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -16,7 +17,14 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
+    {        
+        if (! Gate::allows('isUser', request()->user())) {
+            return response()->json([
+                'status' => 403,
+                'data'   => "You dont have access to this action!"
+            ]);
+        }
+
         return response()->json([
 
             'status' => 200,
@@ -33,17 +41,24 @@ class PostController extends Controller
     public function store(PostRequest $request)
     {
 
-        $post = Post::create( array_merge(
+        if (! Gate::allows('isWrite', request()->user())) {
+            return response()->json([
+                'status' => 403,
+                'data'   => "You dont have permission to write post!"
+            ]);
+        }
 
+        $post = Post::create( array_merge(
+            
             $request->all(), 
             ['user_id' => $request->user()->id]
         ));
-
+        
         return response()->json([
-
+            
             'status'    => 200,
             'post_id'   => $post->id
-        ]);
+            ]);
         
     }
 
@@ -71,7 +86,14 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(PostRequest $request, Post $post)
-    {
+    {        
+        if (! Gate::allows('isEdit', request()->user(), $post)) {
+            return response()->json([
+                'status' => 403,
+                'data'   => "You dont have permission to Edit this post!"
+            ]);
+        }
+
         $params = $request->all();
         $post->update($params);
 
@@ -91,12 +113,17 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-
+        if (! Gate::allows('isDelete', request()->user())) {
+            return response()->json([
+                'status' => 403,
+                'data'   => "Only Admin has permission to Delete posts!"
+            ]);
+        }
         $post->delete();
         return response()->json([
 
             'status'    => 200,
-            'post_id'   => $post
+            'post_id'   => $post->id
         ]);
